ENH: Store and process --call-fmt as a string
The current processing of the cmdexec value has a few limitations.
The replacement only occurs if cmdexec is specified as a list of
arguments.  This prevents specifying commands with redirection,
piping, and so on because those must be passed to run as a string.
Also, the replacement is restricted to items that match a placeholder
exactly, so the placeholder can't be a part of a substring.

Support these things by (1) requiring that cmdexec be a string and (2)
converting the run command to a string it can be used as an argument
to the cmdexec format call.  The conversion of the command to the
string representation is in line with datalad run, which does the same
since 9c3a65e3d (RF: run: Always use a string for the command,
2018-05-29).

Re: https://github.com/datalad/datalad-container/pull/35#discussion_r192040887
Re: https://github.com/datalad/datalad/pull/2597