ENH: add: Copy local files instead of transforming them to a URL
As of git-annex 6.20180626, using a local file as the argument to
--url will fail with the default configuration for
annex.security.allowed-url-schemes.  Adding a local image is primarily
done for testing purposes, and pointing to a local URL doesn't provide
much of a benefit, so just copy it instead.  If a user wants to treat
a local path as a URL, they can still specify it with 'file://'.

Closes #47.