ENH: run: Move the image from inputs to extra_inputs
`containers-run` builds on `run`, augmenting the input list with a
Singularity image.  This means that if a user then puts "{inputs}" in
their command, the image file is included too.  This is almost
certainly surprising to the user because the user has not explicitly
specified the image with --input and is probably not aware that
containers-run adds it.  Avoid this by using run_command's recently
added extra_inputs argument.

Re: datalad/datalad#3038
Closes #38.