ENH: add: Support Singularity's "docker://" URI
Ideally a user is able to use the shub:// scheme so that we resolve a
URL to register with git-annex, but, for images that are only
available from Docker, it is still useful to support Singularity's
"docker://" scheme.

Sadly, this introduces the possibility for confusion between the
"docker://" and "dhub://" prefix, but it seems worth it in order to
support pulling in Docker images via Singularity.