BF: add: Abort when container already exists
It's not possible to pull in a new container from a URL by repeating
the containers-add call.  If the referenced container has changed, the
addurl call fails because the checksum has changed.  And if different
containers-add flags are specified with an existing name (e.g., a
different call format), the configuration is updated with the new
values even though the addurl call fails.

Abort if the user is trying to update a name that already exists.  An
upcoming commit will introduce an --update option to containers-add.