NF: add: Support updating a previous container
To update a container, (1) set parameters to previously configured
values if they are not specified and (2) remove an image before
updating it from a URL (usually through addurl) so that the previous
image doesn't interfere with creating the new one.

Skip the second part if the update is restricted to the call format.
Also don't bother with the second part if a different location
(i.e. --image) is specified for the image.  This conservatively keeps
the image in the working tree, leaving it to the user to remove it if
desired.

Closes #58.