ENH: run: Look for containers in subdatasets only if name has "/"
Depending on the dataset structure, searching through subdatasets may
be expensive.  Let's avoid doing so if the container name doesn't
include a slash.  Due to how containers-list generates names, we know
that a container name from a subdataset must include a slash.

Deciding to recurse based on a slash comes with user-facing changes in
the automatic selection of a container that occurs when no container
name is specified.  If the current dataset has only one container,
it will be executed, regardless of whether the subdatasets have
containers.  And if the only container that is configured for the
dataset and _subdatasets_ is a container within a subdataset, it will
no longer be executed when the caller doesn't specify a container
name.

This is still more expensive than it needs to be: (1) deciding to
recurse based on a slash can lead to false positives because a
container can be referred to be its path, which may contain a slash,
and (2) if the name contains a slash, we recurse _all subdatasets_.
Later we may want to do a more targeted lookup of the container name.

Closes #64.