RF: run: Move "find container" logic to a separate module
This logic is a bit involved and takes up a good chunk of the run
module, obscuring the fact that containers_run is a relatively simple
wrapper around the core run.  Move this logic to a separate module,
and break the different "find" behaviors into separate functions to
improve readability.