ENH: {dspath} to point to the dataset containing the container
Relates to https://github.com/datalad/datalad-container/issues/75 where I
blissfully forgot about the "sandwich" of placeholders where some are
evalutated by datalad_containers (ATM only {img} and {cmd}) and the ones
to be passed into datalad-run should be double-wrapped with {}.

Since {dspath} of the dataset where actual execution to be done is of no
concern for variables at containers-run level, I have decided to expose the
(relative) path to it within {dspath}.  containers-list would return "dspath"
record with the full path to it, and then (similarly to image) it would get
converted to the relative path from current ({pwd} for datalad-run) directory.
I have not tried to "rerun" such runs, but it seems to work at least at the
"containers-run" level.

Motivation:

In my case I want to use a custom "runner" script, which is kept in the same
repository as the containers. So I would like to have a variable I could use to
refer to the top of the dataset which contains containers (which might be
different from the dataset where I actually execute, e.g. when dataset with
containers is a subdataset of the outside container).  It also relates but
orthogonal to the #72 (`containers-run` within a subdirectory), where the
[proposed
patch](https://github.com/datalad/datalad-container/issues/72#issuecomment-472925070)
seems to address the issue for running within subdirectory and on my trial
plays nicely with the change proposed here -- I was able to
containers-run a container from within a subdataset within a
sub-directory of the super-dataset.

This PR ATM comes without a unit-test since first I thought to seek feedback
from datalad-containers/run gurus.