ENH: add: Log start of 'singularity build' call
Building a singularity image can take a while, so we should at least
give some indication of progress.  Ideally we would provide feedback
over the course of the process.  However, processing singularity's
output by giving a callable to log_{stdout,stderr} is tricky
because (1) the output as well as target stream differ substantially
between singularity version 2 and version 3 and (2) 'singularity
build' doesn't seem to adjust its output when it's not connected to a
tty, so it tries to do things like update the progress bars in place.

An alternative would be to dump the output to the terminal with
run(..., log_online=True, log_stdout=False, log_stderr), but datalad
generally avoids straight dumps of output from other programs, so
let's stick with an initial log for now.

Re: #69