MNT: Drop Docker adapter
It was a useful idea to explore, but it's unlikely that anyone uses
it.  When an image is only available from Docker, we now support the
"docker://" prefix for building Singularity images, so it's not clear
that this is worth maintaining.  (And, although not the adapter's
fault, its existence sadly holds up the merge of gh-60.)  If there is
something that "docker://" won't work for, we can always resurrect
this.