BF: report "Update " in commit message only if an existing image was actually updated
I tend to use --update while adding or updating existing image.  That leads to
IMHO erroneous commit message that an image was updated whenever it was freshly
added/configured.  Docstring for --update says "If a container with `name` does
not already exist, this option is ignored.", so reporting Conffigure in the
cases when it was "ignored" is IMHO a correct thing to do