ENH: Specify result handling when listing subdataset containers
When we call containers_list() recursively to get the containers in
subdatasets, we leave the result parameters at their default values,
but we should adjust the parameters for these internal calls.  In
particular, on_failure is important so that we let failures bubble up
so that we can react to failures as specified by the caller.

Disabling the result_renderer will be necessary to avoid double
entries when we add a custom renderer.