BF: docker adapter: Fix call to logging.exception()
logging.exception() expects a message, not the exception instance.
Passing the instance pollutes the original traceback output with the
traceback for an unrelated exception.

Note that for consistency with DataLad, this should use exc_str(), but
the docker adapter is written to be independent of DataLad (at least
for now).