TST: docker adapter: Add simple tests
The docker adapter hasn't seen any attention since it was added as a
simple prototype.  If we're going to keep it around and flesh it out,
there are various rough edges that need to be sorted out.  These tests
don't touch on any of those rough edges, but they provide a base for
later series to work with.

One tricky aspect of these tests is that we are using the system's
docker, so we try to leave it in the same state that we found it.
This means that, if there is already a busybox:latest, we don't pull
down a new one and don't delete it after the tests.  Eventually we
should probably create a dedicated image for these tests rather than
using busybox.