ENH: add: Let DataLad handle shub:// when capable
containers-add resolves shub:// URLs to HTTP URLs by querying
<https://singularity-hub.org/api/container/> and then feeds the result
to `git annex addurl`.  That works to download the file, but the
registered URL is short-lived, so it's not useful for retrieving the
image later with `{datalad,git annex} get`.

DataLad's 33429168c (NF: downloaders: Add support for shub:// URLs,
2020-08-11) added a downloader for shub:// URLs.  With the datalad
special remote enabled, shub:// URLs can be registered directly with
git-annex, and the URL will be resolved at the time of the
`{datalad,git annex} get` call.

If the shub downloader is detected, avoid the custom handling on
containers-add's end.

Closes #96.