RF: find_container: Convert operation to a generator
An upcoming commit will add logic that installs necessary subdatasets.
In that case, ideally containers_run() would be able to yield the
result record for the install actions.  Move the main logic of
find_container() into a generator to prepare for this.

Keep the non-generator of find_container() around since it's already
used in third-party code (reproman).