ENH: run: Try to install container subdataset if needed
When a container name with a slash is given, containers_run() (via
find_container.py) checks the name against those that
containers_list(..., recursive=True) reports.  In order for a
subdataset container to be known, the subdataset must already be
installed.

To lift this requirement, teach find_container_() to install the
corresponding subdataset path for each component in the
container_name.  The operation isn't as simple as a `get <leading path
of name>` because the submodule's name doesn't necessary match its
path; instead submodule inspection and a get() call is needed for each
dataset.

The operation is also complicated by the fact that container_name
isn't always a chain of module names: containers_run() also accepts an
image name as the container_name.  This means that a slash in the name
doesn't always map to a subdataset boundary.

Closes #122.