TST: utils: Add helper to set up simple script as image
The tests use two types of images: (1) proper images that can be
executed and (2) dummy files that can't actually run but are useful
for lighter testing of logic outside containers_run().  Add something
that's in between: a local script that's not a proper container but
can be executed with containers_run().