TST: Adjust more tests for WitlessRunner-based run()
datalad/datalad#5002 switches `datalad run` over to using
WitlessRunner rather than the old Runner.  swallow_outputs() works for
Runner because it explicitly sets the stdout argument of Popen() to
sys.stdout, which allows the swallow_outputs() override to be
effective.  This is not the case for WitlessRunner.

A few commits back already adjusted
datalad_container/tests/test_run.py for the above change in behavior.
Use the same approach in the remaining spots that swallow and check
the output of a containers_run() call.

https://github.com/datalad/datalad/pull/4996#pullrequestreview-505761685