Prototype of script to feed Docker Hub tags to containers-add
It's rough, but it might be at least somewhat functional.  I let the
first few entries of

    echo repronim/ | python containers_add_dhub_tags.py

and

    echo neurodebian | python containers_add_dhub_tags.py

complete.  The lack of progress about the underlying `docker pull` is
unfortunate (and I bet there's a -container issue open about it).

The main design decision here is to name the results (the image and
downloaded manifest) based on the manifest's .config.digest value and
then use exists() checks as an indication that we already have the
result locally.  Whether that's actually valid should be revisited.

Ref: https://github.com/ReproNim/containers/issues/48