Use tag in containers-add name
The digest was used in the first pass to avoid worrying about invalid
characters, but as mentioned in the comment and on the
ReproNim/containers issue tracker [*], this isn't good for recognizing
the name.

Instead construct the name from combining the repository and tag,
replacing any characters that containers-add doesn't allow with "--".
This introduces ambiguity and the potential for conflicts but is
probably good enough.

[*] https://github.com/ReproNim/containers/issues/48#issuecomment-719135394