ENH: oci: Copy over tag when saving OCI directory
An image stored as an OCI directory can have a tag.  If the source has
a tag specified, copy it over to the destination.

Note that in upcoming commits will store the full source specification
as an image annotation, so we won't rely on this when copying the
image to docker-daemon:, but it still seems nice to have (e.g., when
looking at the directory with skopeo-inspect).