NF: oci: Add adapter for working with OCI images
In order to be able to track Docker containers in a dataset, we
introduced the docker-save-based docker adapter in 68a1462 (Add
prototype of a Docker adapter, 2018-05-18).  It's not clear how much
this has been used, but at least conceptually it seems to be viable.
One problem, however, is that ideally we'd be able to assign Docker
registry URLs to the image files stored in the dataset (particularly
the large non-configuration files).  There doesn't seem to be a way to
do this with the docker-save archives.

Another option for storing the image in a dataset is the Open
Container Initiative image format.  Skopeo can be used to copy images
in Docker registries (and some other destinations) to an OCI-compliant
directory.  When Docker Hub is used as the source, the resulting
layers blobs can be re-obtained via GET /v2/NAME/blobs/ID.

Using skopeo/OCI also has the advantage of making it easier to execute
via podman in the future.

Add an initial skopeo-based OCI adapter.  At this point, it has the
same functionality as the docker adapter.