ENH: containers-add: Wire up OCI adapter
Add a new oci: scheme.  The stacking of the schemes isn't ideal
(oci:docker://, oci:docker-daemon:), but it allows for any skopeo
transport to be used.

Note: I'm not avoiding appending "//" for a conceptual reason
(although there might be a valid one), but because I find
"oci://docker://" to be ugly.  Perhaps the consistency with "shub://"
and "dhub://" outweighs that though.