ENH: oci: Silence 'skopeo copy' output when loading image
Prevent skopeo-copy output from being shown, since it's probably
confusing to see output under run's "Command start (output follows)"
tag for a command that the user didn't explicitly call.  However, for
large images, this has the downside that the user might want some
signs of life, so this may need to be revisited.