ENH: oci: Register a more informative name with docker-daemon
The images copied to the daemon look like this

    $ docker images
    REPOSITORY             TAG                 IMAGE ID            CREATED             SIZE
    datalad-container/bb   sha256-98345e4      98345e418eb7        3 weeks ago         69.2MB

That tag isn't useful because it just repeats the image ID.  And the
name after "datalad-container/" is the name of the directory, so with
the default containers-add location it would be an uninformative
"image".

With the last commit, we store the source specification as an
annotation in the OCI directory.  Parse it and reuse the original
repository name and tag.

   REPOSITORY                 TAG                 IMAGE ID            CREATED             SIZE
   datalad-container/debian   buster-slim         98345e418eb7        3 weeks ago         69.2MB

If the source has a digest instead of the tag, construct the daemon
tag from that.