RF: adapters: Prefer subprocess.run over check_* methods
The minimum Python version of DataLad is new enough that we can assume
subprocess.run() is available.  It's recommended by the docs, and I
like it more, so switch to it.

Note that we might want to eventually switch to using WitlessRunner
here.  The original idea with using the subprocess module directly was
that it'd be nice for the docker adapter to be standalone, as nothing
in the adapter depended on datalad at the time.  That's not the case
anymore after the adapters.utils split and the use of datalad.utils
within it.  (And the upcoming skopeo adapter will make heavier use of
datalad for adding URLs to the layers.)