ENH: adapters.utils: Account for datalad logging
When the adapter is called from the command line (as containers-run
does) and datalad gets imported, the level set via the --verbose
argument doesn't have an effect and logging happens twice, once
through datalad's handler and once through the adapter's.

Before 313c4f0f (WIN/Workaround: don't pass gid and uid to docker run
call, 2020-11-10), the above was the case when docker.main() was
triggered with the documented `python -m datalad_container.adapters
...` invocation, but not when the script path was passed to python.
Following that commit, the adapter imports datalad, so datalad's
logger is always configured.

Adjust setup_logger() to set the log level of loggers under the
datalad.containers.adapters namespace so that the adapter's logging
level is in effect for command line calls to the adapter.

As mentioned above, datalad is now loaded in all cases, so a handler
is always configured, but, in case this changes in the future, add a
simpler handler if one isn't already configured.