Slightly relax tests to account for upcoming remove() change
Previously, removing a file could report the removed path in
relative form, unlike any other datalad command. The replacement
in https://github.com/datalad/datalad/pull/6111 unifies this.
The test is changed to no rely on the path format -- which is
just a footnote re the purpose of the test.