RM: remove trying to provide cont_path to container image config
Because then we would need to format in other places like for get etc,
not to mention that likely in reproman would need to do more magic.
Since it makes sense ATM only to have path from the location where container
is defined, just keep it that way and not bother interpolating.