Make matrix with different python versions and also separate one where we do pre-remove busybox (on linux only)
I believe some coverage dropage while removing travis config is because there we did matrix across many python versions
but also what is important, we had travis and appveyor contributing. I guess on travis we did not have that busybox
pre-fetched and on appveyor we did have it prefetched. As a result we had both scenarios.  Hopefully we regain more of
coverage here now