RF `containers_remove` to not need a `remove` call
The previous implementation only removed the handles for an image
and did not actually drop anything.

This is done (simpler), but just removing the files on the filesystem,
and having `save()` commit the change.

This is faster, because `save()` needs to run anyways. And cleaner
(the previous implementation did a no-drop-remove, which it the exact
same thing, but runs a complex software stack to do that).

Moreoever, this changeset discontinues the use of a deprecated API
aspect of `remove`, by not using `remove` at all.

Removal of individual files is still announced via individual results.

Closes #240